requirements.txt
# 🍖 Nepal Meat Shop - Dependencies
# Core Flask framework and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5

# Database support
SQLAlchemy==2.0.23

# MongoDB support
pymongo==4.6.0
flask-pymongo==2.3.0

# Form handling and validation
WTForms==3.1.1

# Security and password hashing
Werkzeug==3.0.1

# File upload handling
Pillow==10.1.0

# PDF generation
reportlab==4.0.7

# Environment configuration
python-dotenv==1.0.0

# Development and testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Production server (optional)
gunicorn==21.2.0